Attempt gcloud logging
