+++ shared/util.go
@@ -16,7 +16,6 @@ import (
 	log "github.com/Hexcles/logrus"
 	mapset "github.com/deckarep/golang-set"
 	"google.golang.org/appengine"
-	gaelog "google.golang.org/appengine/log"
 )
 
 // ExperimentalLabel is the implicit label present for runs marked 'experimental'.
@@ -94,125 +93,6 @@ func ToStringSlice(set mapset.Set) []string {
 // of which are treated as looking up the latest run for each browser.
 func IsLatest(sha string) bool {
 	return sha == "" || sha == "latest"
-}
-
-// Logger is an abstract logging interface that contains an intersection of
-// logrus and GAE logging functionality.
-type Logger interface {
-	Debugf(format string, args ...interface{})
-	Errorf(format string, args ...interface{})
-	Infof(format string, args ...interface{})
-	Warningf(format string, args ...interface{})
-}
-
-// SplitLogger is a logger that sends logging operations to both A and B.
-type loggerMux struct {
-	delegates []Logger
-}
-
-type gaeLogger struct {
-	ctx context.Context
-}
-
-type nilLogger struct{}
-
-// Debugf implements formatted debug logging to both A and B.
-func (lm loggerMux) Debugf(format string, args ...interface{}) {
-	for _, l := range lm.delegates {
-		l.Debugf(format, args...)
-	}
-}
-
-// Errorf implements formatted error logging to both A and B.
-func (lm loggerMux) Errorf(format string, args ...interface{}) {
-	for _, l := range lm.delegates {
-		l.Errorf(format, args...)
-	}
-}
-
-// Infof implements formatted info logging to both A and B.
-func (lm loggerMux) Infof(format string, args ...interface{}) {
-	for _, l := range lm.delegates {
-		l.Infof(format, args...)
-	}
-}
-
-// Warningf implements formatted warning logging to both A and B.
-func (lm loggerMux) Warningf(format string, args ...interface{}) {
-	for _, l := range lm.delegates {
-		l.Warningf(format, args...)
-	}
-}
-
-func (l gaeLogger) Debugf(format string, args ...interface{}) {
-	gaelog.Debugf(l.ctx, format, args...)
-}
-
-func (l gaeLogger) Errorf(format string, args ...interface{}) {
-	gaelog.Errorf(l.ctx, format, args...)
-}
-
-func (l gaeLogger) Infof(format string, args ...interface{}) {
-	gaelog.Infof(l.ctx, format, args...)
-}
-
-func (l gaeLogger) Warningf(format string, args ...interface{}) {
-	gaelog.Warningf(l.ctx, format, args...)
-}
-
-func (l nilLogger) Debugf(format string, args ...interface{}) {}
-
-func (l nilLogger) Errorf(format string, args ...interface{}) {}
-
-func (l nilLogger) Infof(format string, args ...interface{}) {}
-
-func (l nilLogger) Warningf(format string, args ...interface{}) {}
-
-// LoggerCtxKey is a key for attaching a Logger to a context.Context.
-type LoggerCtxKey struct{}
-
-var (
-	gl  = gaeLogger{}
-	nl  = nilLogger{}
-	lck = LoggerCtxKey{}
-)
-
-// NewLoggerMux creates a multiplexing Logger that writes all log operations to
-// all delegates.
-func NewLoggerMux(delegates []Logger) Logger {
-	if len(delegates) == 0 {
-		return NewNilLogger()
-	}
-	return loggerMux{delegates}
-}
-
-// NewGAELogger returns a Google App Engine Standard Environment logger bound to
-// the given context.
-func NewGAELogger(ctx context.Context) Logger {
-	return gaeLogger{ctx}
-}
-
-// NewNilLogger returns a new logger that silently ignores all Logger calls.
-func NewNilLogger() Logger {
-	return nl
-}
-
-// DefaultLoggerCtxKey returns the default key where a logger instance should be
-// stored in a context.Context object.
-func DefaultLoggerCtxKey() LoggerCtxKey {
-	return lck
-}
-
-// GetLogger retrieves a non-nil Logger that is appropriate for use in ctx. If
-// ctx does not provide a logger, then a nil-logger is returned.
-func GetLogger(ctx context.Context) Logger {
-	logger, ok := ctx.Value(DefaultLoggerCtxKey()).(Logger)
-	if !ok || logger == nil {
-		log.Warningf("Context without logger: %v; logs will be dropped", ctx)
-		return NewNilLogger()
-	}
-
-	return logger
 }
 
 // NewAppEngineContext creates a new Google App Engine Standard-based
